feat: 添加 DSR/ITEA/MRGP 回归器及符号表达式安全解析
新增三个符号回归算法的适配器模块，扩展 evaluation.py
中的符号解析器以支持 sub/div/square/cube/PLOG/PSQRT 等
安全数学算子，修复 FEAT 模型提取接口，并在 analyze.py
中增加 n_jobs 传递和噪声结果跳过逻辑。

+++ experiment/evaluate_model.py
@@ -124,7 +124,7 @@ def evaluate_model(
         sc_y = StandardScaler()
         y_train_scaled = sc_y.fit_transform(y_train.reshape(-1,1)).flatten()
     else:
-        y_train_scaled = y_train
+        y_train_scaled = y_train.astype(float)
 
 
     ##################################################
